db: separate database module into session and engine
Separate database module into session and engine modules.
This way the engine can be used by alembic later.
